+++ src/setupSigintHandler.ts
@@ -1,6 +1,0 @@
-export const setupSigintHandler = (stopCallback: () => Promise<void>) => {
-  process.on("SIGINT", async () => {
-    await stopCallback();
-    process.exit(0);
-  });
-};